pyproject.toml
[build-system]
  build-backend = "hatchling.build"
  requires = ["hatchling"]

[project]
  authors = [{ name = "Paul Razvan Berg" }]
  dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
  ]
  description = "CLI tool for parsing and processing Claude Code prompt flags"
  name = "ai-flags"
  readme = "README.md"
  requires-python = ">=3.12"
  version = "0.1.0"

[project.optional-dependencies]
  dev = [
    "click>=8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyright>=1.1",
    "pytest>=8.0",
    "pytest-mock>=3.0",
    "pyyaml>=6.0",
    "ruff>=0.8",
  ]

[project.scripts]
  ai-flags = "ai_flags.cli:cli"

[tool.hatch.build.targets.wheel]
  packages = ["src/ai_flags"]

[tool.pytest.ini_options]
  pythonpath = "src"
  testpaths = ["tests"]
